pyproject.toml
[project]
name = "neoparatranz"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "lxml>=6.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "datamodel-code-generator[http]>=0.31.2",
    "nuitka>=2.7.11",
]

